initial refactor for an example
