+++ health/src/main/java/uk/gov/defra/tracesx/common/health/Check.java
@@ -1,0 +1,8 @@
+package uk.gov.defra.tracesx.common.health;
+
+import org.springframework.boot.actuate.health.Health;
+
+public interface Check {
+
+    Health check();
+}
